refactor:organizationHelper.ts
