+++ app/api/api-constants.ts
@@ -1,0 +1,2 @@
+export const ORG_PROFILE = '/orgs/public-profile'
+export const ORG_DETAILS = '/orgs/public-profiles/'
