+++ app/constants.ts
@@ -75,8 +75,6 @@ export const PINRules: PINValidationRules = {
 }
 
 export const domain = 'didcomm://invite'
-export const ORG_PROFILE = '/orgs/public-profile'
-export const ORG_DETAILS = '/orgs/public-profiles/'
 
 export const tourMargin = 25
