+++ app/screens/OrganizationDetails.tsx
@@ -4,10 +4,10 @@ import { useTranslation } from 'react-i18next'
 import { View, StyleSheet, Text, Image, ScrollView } from 'react-native'
 import Toast from 'react-native-toast-message'
 
+import useOrganizationDetailData from '../api/organizationDetailHelper'
 import Button, { ButtonType } from '../components/buttons/Button'
 import { ToastType } from '../components/toast/BaseToast'
 import { useTheme } from '../contexts/theme'
-import useOrganizationDetailData from '../hooks/organizationDetailHelper'
 import { BifoldError } from '../types/error'
 import { Screens, Stacks } from '../types/navigators'
 import { useAppAgent } from '../utils/agent'
